chore(activists): add detail render #815
