+++ routes/admin/authenticated/sidebar/activists/container.js
@@ -1,97 +1,214 @@
 import React, { Component } from 'react'
-import { FormattedMessage } from 'react-intl'
 import {
   SettingsPageLayout,
   SettingsPageMenuLayout,
   SettingsPageContentLayout
 } from '~client/components/layout'
-
-import PaginationGrid from './pagination-grid'
+import { Loading } from '~client/components/await'
 
 if (require('exenv').canUseDOM) require('./styles.scss')
 
-// move component
-const ActivistDetail = ({ item }) => (
-  <ul>
-    <li>Nome: {item.name}</li>
-    <li>E-mail: {item.email}</li>
-    <li>Telefone: {item.phone}</li>
-  </ul>
+const QueryForm = ({
+  query,
+  onQueryChange,
+  onSubmit,
+  label,
+  buttonText,
+  name
+}) => (
+  <form
+    className='query-form pr4 pl3 border-box'
+    onSubmit={e => {
+      e.preventDefault()
+      onSubmit()
+    }}
+  >
+    <i className='fa fa-search' aria-hidden='true'></i> 
+    <input
+      id={`${name}Id`}
+      type='text'
+      onChange={onQueryChange}
+      value={query}
+      placeholder={label}
+    />
+    {buttonText && (
+      <input
+        className='absolute'
+        type='submit'
+        value={buttonText}
+      />)}
+  </form>
 )
-// end todo
 
-export default class ActivistsContainer extends Component {
-  
+const Pagination = ({
+  indexPage,
+  lastPage,
+  onPreviousPage,
+  onNextPage
+}) => {
+  const handleClick = (handle) => (e) => {
+    e.preventDefault()
+    handle()
+  }
+  return (
+    <div className='pagination clearfix mx-auto'>
+      <div className='col col-1'>
+        <button
+          disabled={indexPage < 2}
+          onClick={handleClick(onPreviousPage)}
+        >
+          <i className='fa fa-arrow-left' aria-hidden="true"></i>
+        </button>
+      </div>
+      <div className='col col-10'>
+        <p>{indexPage} / {lastPage}</p>
+      </div>
+      <div className='col col-1'>
+        <button
+          disabled={indexPage === lastPage}
+          onClick={handleClick(onNextPage)}
+        >
+          <i className='fa fa-arrow-right' aria-hidden='true'></i>
+        </button>
+      </div>
+    </div>
+  )
+}
+
+const Row = ({ obj, onSelectRow, onClickRow, isSelected, isActived }) => (
+  <div
+    className={`row clearfix mx-auto p2${isActived ? ' active' : ''}`}
+    onClick={() => onClickRow(obj)}
+  >
+    <div className='col col-1'>
+      <input
+        type='checkbox'
+        checked={isSelected}
+        onChange={() => onSelectRow(obj.id)}
+      />
+    </div>
+    <div className='col col-6'>
+      {obj.name}
+    </div>
+    <div className='col col-5'>
+      {obj.email}
+    </div>
+  </div>
+)
+
+
+const Detail = ({ obj, onClose }) => (
+  <div className='detail col col-4 pl2'>
+    <div className='col col-12 title py2'>
+      <span>Perfil</span>
+      <i
+        className='fa fa-close'
+        aria-hidden='true'
+        onClick={() => onClose(obj)}
+      />
+    </div>
+    <div className='col col-12 bg-white py1 px2'>
+      <h3>{obj.name}</h3>
+      <div>
+        <div>
+          <label>E-mail</label>
+          <p>{obj.email}</p>
+        </div>
+      </div>
+    </div>
+  </div>
+)
+
+
+class Container extends Component {
+
   constructor (props) {
     super(props)
-    this.state = {
-      item: null,
-      rowIndex: null,
-      rowIndexList: [],
-      query: '',
-    }
+    this.state = { item: null }
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.location &&
-      nextProps.location.query &&
-      nextProps.location.query.q) {
-      
-      this.setState({ query: nextProps.location.query.q })
-    }
+  renderDetail () {
+    return (
+      <div className='col col-3'>
+      </div>
+    )
   }
 
-  handleClickRow (item, rowIndex) {
-    if (this.state.rowIndex === rowIndex) {
-      this.setState({ rowIndex: null, item: null })
+  onClickRow (item) {
+    if (this.state.item === item) {
+      this.setState({ item: null })
     } else {
-      this.setState({ rowIndex, item })
+      this.setState({ item })
     }
-  }
-
-  onQueryChange (e) {
-    this.setState({ query: e.target.value })
-  }
-
-  onQuerySubmit (e) {
-    e.preventDefault()
-    const { communityId, refetch } = this.props
-    refetch({ communityId, search: this.state.query })
   }
 
   render () {
     
-    const { totalCount } = this.props
+    const { data, loading, totalCount } = this.props
+    const { indexPage, lastPage, onNextPage, onPreviousPage } = this.props
+
+    const { selected, onSelectRow, onSelectAll, selecting } = this.props
 
     return (
       <SettingsPageLayout>
-        <SettingsPageMenuLayout
-          title={
-            <FormattedMessage
-              id='activists.routes--container.title'
-              defaultMessage='Base de usuários'
-            />
-          }
+        <SettingsPageMenuLayout title='Base de usuários' />
+        <QueryForm
+          name='q'
+          label='Filtre por mobilizações ou formulários'
+          onSubmit={this.props.fetch}
+          query={this.props.query}
+          onQueryChange={this.props.onQueryChange}
         />
         <SettingsPageContentLayout>
-          <form onSubmit={this.onQuerySubmit.bind(this)}>
-            <input
-              type='text'
-              name='q'
-              onChange={this.onQueryChange.bind(this)}
-              value={this.state.query}
-            />
-            <button type='submit'>Buscar</button>
-          </form>
-          <h3>{totalCount} ativistas encontrados</h3>
-          <div>
-            {this.state.item && (
-              <ActivistDetail item={this.state.item} />
-            )}
-            <PaginationGrid {...this.props} />
+          <div className={(
+            this.state.item ? 'col col-8' : ''
+          )}>
+            <div className='title clearfix p2'>
+              <div className='col col-1'>
+                <input
+                  id='selectAllId'
+                  type='checkbox'
+                  checked={selected.length > 0 && selected.length === totalCount}
+                  onClick={(evt) => {
+                    onSelectAll()
+                  }}
+                />
+              </div>
+              <div className='col col-11'>
+                {totalCount} pessoas
+              </div>
+            </div>
+            {(selecting || loading) && <Loading />}
+            <div>
+              <div>
+                {data.map(d => (
+                  <Row
+                    obj={d}
+                    isSelected={selected.indexOf(d.id) !== -1}
+                    isActived={this.state.item === d}
+                    onSelectRow={this.props.onSelectRow}
+                    onClickRow={this.onClickRow.bind(this)}
+                  />
+                ))}
+              </div>
+              <Pagination
+                indexPage={this.props.indexPage}
+                lastPage={this.props.lastPage}
+                onPreviousPage={this.props.onPreviousPage}
+                onNextPage={this.props.onNextPage}
+              />
+            </div>
           </div>
+          {this.state.item && (
+            <Detail
+              obj={this.state.item}
+              onClose={this.onClickRow.bind(this)}
+            />
+          )}
         </SettingsPageContentLayout>
       </SettingsPageLayout>
     )
   }
 }
+
+export default Container
