+++ routes/admin/authenticated/sidebar/activists/styles.scss
@@ -26,22 +26,28 @@
   }
 }
 
-.counter {
+.title {
+  position: relative;
   font-family: 'Source Sans Pro', 'Proxima Nova', sans-serif;
   font-size: 35px;
   color: #222;
-  padding: 16px;
+
+  i {
+    position: absolute;
+    cursor: pointer;
+    right: 0;
+    top: 18px;
+  }
 }
 
 .row {
   font-family: 'Source Sans Pro', 'Proxima Nova', sans-serif;
   font-size: 14px;
   color: #333;
-  padding: 16px 17px;
   background-color: #fff;
   border-bottom: solid 1px #efefef;
 
-  &:hover {
+  &:hover, &.active {
     cursor: pointer;
     border-left: solid 3px #ff0099;
     padding-left: 14px;
@@ -70,5 +76,15 @@
   }
   button[disabled] {
     color: #c7c7c7;
+  }
+}
+
+.detail {
+    
+  h3 {
+    font-family: 'Source Sans Pro', 'Proxima Nova', sans-serif;
+    font-size: 22px;
+    font-weight: bold;
+    color: #222222;
   }
 }
